+++ mandelbulber2/src/opencl_engine_render_dof_phase2.h
@@ -60,8 +60,8 @@ public:
 	void SetParameters(const sParamRender *paramRender, const cRegion<int> &region);
 	bool LoadSourcesAndCompile(const cParameterContainer *params) override;
 	void RegisterInputOutputBuffers(const cParameterContainer *params) override;
-	bool AssignParametersToKernelAdditional(int argIterator, int deviceIndex) override;
-	bool ProcessQueue(qint64 pixelsLeft, qint64 pixelIndex);
+	bool AssignParametersToKernelAdditional(uint argIterator, int deviceIndex) override;
+	bool ProcessQueue(quint64 pixelsLeft, quint64 pixelIndex);
 	bool Render(cImage *image, cPostRenderingDOF::sSortZ<float> *sortedZBuffer, bool *stopRequest);
 	size_t CalcNeededMemory() override;
 
@@ -71,7 +71,7 @@ private:
 	const int outputIndex = 0;
 	QString GetKernelName() override;
 	sParamsDOF paramsDOF;
-	int numberOfPixels;
+	quint64 numberOfPixels;
 	cRegion<int> imageRegion;
 
 #endif
